feat: Update Norwegian translation for "thankYou" message
This commit updates the Norwegian translation for the "thankYou" message in the translation file. The previous translation was "Takk for at du besøkte!" and it has been changed to "Takk for besøket!". This change improves the clarity and simplicity of the message, providing a better user experience for Norwegian-speaking visitors.